+++ packages/appkit/CHANGELOG.md
@@ -12,7 +12,7 @@
   - @rerock/appkit-common@0.0.3
   - @rerock/appkit-core@0.0.3
   - @rerock/appkit-polyfills@0.0.3
-  - @rerock/scaffold-ui@0.0.3
+  - @rerock/appkit-scaffold-ui@0.0.3
   - @rerock/scaffold-utils@0.0.3
   - @rerock/siwe@0.0.3
   - @rerock/ui@0.0.3